Lesson-1 Исправление замечаний проверки
