+++ lesson-1/README.md
@@ -1,6 +1,6 @@
 ﻿OTUS Spring course homeworks
 
-Островский Дмитрий (Ostrovskiy Dmitriy) Spring Framework для Diasoft Группа 2021-03
+Соловьянова Наталья (Solovyanova Natalya) Spring Framework для Diasoft Группа 2021-03
 
 Домашнее задание Программа по проведению викторины. • В ресурсах хранятся вопросы и различные ответы к ним в виде CSV файла. • Программа должна спросить у участника имя и возраст спросить вопросы из CSV-файла и вывести результат викторины.
