+++ lesson-1/src/main/java/ru/otus/spring/Main.java
@@ -1,18 +1,15 @@
 package ru.otus.spring;
 
-//import org.springframework.context.support.ClassPathXmlApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
-import ru.otus.spring.domain.Question;
 import ru.otus.spring.service.QuestionService;
 
 public class Main {
 
     public static void main(String[] args) throws Exception {
-        // TODO: создайте здесь класс контекста
         ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext("/spring-context.xml");
 
         QuestionService service = context.getBean(QuestionService.class);
         service.runTest();
-        context.close();;
+        context.close();
     }
 }
