+++ lesson-1/src/main/java/ru/otus/spring/dao/ParserCsvDao.java
@@ -1,0 +1,9 @@
+package ru.otus.spring.dao;
+
+import ru.otus.spring.domain.Question;
+
+import java.io.InputStream;
+
+public interface ParserCsvDao {
+    public Question getParseQuestionsFromCsv(InputStream input);
+}
