+++ lesson-1/src/main/java/ru/otus/spring/dao/QuestionDao.java
@@ -3,5 +3,5 @@ package ru.otus.spring.dao;
 import ru.otus.spring.domain.Question;
 
 public interface QuestionDao {
-    public Question getQuestionsFromCsv();
+    Question getQuestions();
 }
