+++ lesson-1/src/test/java/ru/otus/spring/dao/QuestionDaoImplTest.java
@@ -2,7 +2,6 @@ package ru.otus.spring.dao;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import ru.otus.spring.dao.QuestionDaoSimple;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -10,14 +9,15 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 @DisplayName("Вопросы для теста")
 public class QuestionDaoImplTest {
     private QuestionDaoSimple questions;
+    private ParserCsvDao parse;
 
     @Test
     @DisplayName("Генерим исключение, если файла не существует")
     void shouldIOExceptionWhenFileNotFound() {
         String questionFile = "test.csv";
         Exception exception = assertThrows(Exception.class, () -> {
-            questions = new QuestionDaoSimple(questionFile);
-            questions.getQuestionsFromCsv();
+            parse = new ParserCsvDaoSimple();
+            questions.getQuestions();
         });
 
         String expectedException = new StringBuilder("File not found ")
